Remove the crypto library from gameUtils
This makes the library easier to use in the browser. Technically, the
game's random state is a wee bit more "guessable" (if you can guess the
millisecond the game started on, and you have a mersenne-twister handy.
However, this should be fine, and if we _REALLY_ need more entropy
during game init, we can use Math.random() * 2**32 to init the MT
library, since most browsers / JS runtimes will try to seed the native
RNG with a good OS-level source.